Change tests for the fragment
